chore(profiling): Split up flamegraph state file
This splits up the flamegraph state into multiple files to make it easier to
follow as more is added to the state.
